refactor(alert-manager): apply changed router
