+++ apps/web/src/services/alert-manager-v1/components/EscalationPolicyDataTable.vue
@@ -15,11 +15,11 @@ import type {
 } from '@/schema/monitoring/escalation-policy/type';
 import { i18n } from '@/translations';
 
+import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
 import { useAllReferenceStore } from '@/store/reference/all-reference-store';
 
 import { referenceRouter } from '@/lib/reference/referenceRouter';
 
-import { useProperRouteLocation } from '@/common/composables/proper-route-location';
 import { useProxyValue } from '@/common/composables/proxy-state';
 
 import { alertStateBadgeStyleTypeFormatter, alertResourceGroupBadgeStyleTypeFormatter } from '@/services/alert-manager-v1/helpers/alert-badge-helper';
@@ -40,9 +40,9 @@ const props = withDefaults(defineProps<{
 const emit = defineEmits<{(e: 'update:select-index', value: number[]): void;
     (e: 'change', value: { sortBy: string; sortDesc?: boolean }): void;
 }>();
-const { getProperRouteLocation } = useProperRouteLocation();
 
 const allReferenceStore = useAllReferenceStore();
+const userWorkspaceStore = useUserWorkspaceStore();
 const state = reactive({
     projects: computed(() => allReferenceStore.getters.project),
     finishConditions: computed<Record<EscalationPolicyFinishCondition, TranslateResult>>(() => ({
@@ -135,9 +135,12 @@ const onChangeSort = (sortBy: string, sortDesc?: boolean) => {
             <template v-if="value && value !== '*'">
                 <p-link :action-icon="ACTION_ICON.INTERNAL_LINK"
                         new-tab
-                        :to="getProperRouteLocation(referenceRouter(
+                        :to="referenceRouter(
                             value,
-                            { resource_type: 'identity.Project' }))"
+                            {
+                                resource_type: 'identity.Project',
+                                workspace_id: userWorkspaceStore.getters.currentWorkspaceId,
+                            })"
                 >
                     {{ state.projects[value] ? state.projects[value].label : value }}
                 </p-link>
